Map and Tile Changes
Alright, so there's a lot going on with the map and the tiles, so
this'll be a long description.

First, I created Map2 and Tile2, so I would not have to worry about
accidentally deleting anyone's work. Other than that, they should do
everything that's already in place, in case other code depends on it. A
method or two may need to be renamed.

For Tile2, each tile holds three things: in slot 0 of the array; a base
tile (ground, ocean, mountain). In slot 1 of the array; a resource tile
(mineral, methane, noresource). In slot 2 of the array; a weather tile
(noweather, sun). I did an array instead of a stack, because the stack
seemed a bit excessive.

Now, the actual tiles, like Ground or MethaneTile extend the appropriate
class for their role in the array, which extends TileWrapper. I could
split these up, so it hold individual Ground or MethaneTile objects, not
in an array, but there are enough redundant methods among them I kept
them extending the class.

The Map2 creates a Tile2 map, using the diamond-square algorithm I
demoed. It does not have the level of depth of terrain and resource
generation I would like, but it is functional. Likewise, actual
resources (as in the kind that will be used to make objects) are not at
all implemented, the classes just exist for being put into a
ResourceTile. Further, weather does not currently do anything; it is a
placeholder for now.