Improved Tile
The logic behind each Tile is now improved; I slightly changed map
generation so that we have four types of terrain (ocean, sand, plain,
mountain) and three resources placed (methane, carbon, iron). Currently,
all mountain have iron, sand has carbon, and ocean has methane. This is
easy to change.

At the moment, I have set it up such that either the ground or the
building on it determines the text representation, any resource on it
determines color, and weather determines the brightness.

The graphical view, later, will just be each layer of the tile drawn in
sequence, but that doesn't matter now.

No functionality is currently lost, but we will need to implement
building functionality. I'll be working on the methods for adding
buildings/building tiles to the map shortly.