Map, Tile, and Building Updates
-Map now allows us to set a particular tile, given the kind of tile we
want to place, the x position, and the y position

-Map allows us to place a building, filling up those spots on the map
with tiles holding the appropriate building tile. It does not yet add
buildings to a list; this still needs to be done.

-Building has been updated to work with the current map/tile classes.
Primarily, the method to check if a building can be built has been
updated and commented a bit further, with what the code seemed to be
doing to me. Additionally, a building now holds the building enum to
tell the map what to place.

-Base and SolarPlant have had building enums added, to follow the
general Building

-GroundTile has been updated with another getter.

-BuildingEnum had a BASE enum added to it.

Significant tasks remaining:
-add the list of buildings to the map/game loop? (do drones still need
to be added?)
--check how buildings execute each game loop; I don't know off the top
of my head right now
-add something to manage our pool of resources
-UI for issuing commands
-Others that I am forgetting right now