Comment About Map Sizes
As we currently have it, each tile on the map is a separate instance, so
that they can be nodes. This has nearly put us at the limit of memory
Java can use. Currently, the map is 2^7*2^7. 2^8 cannot run, due to
memory issues.

We must either accept this is the largest map possible, or cut some
memory usage somewhere. Probably would have to be from the map, ie,
don't treat a tile as a node and figure out some other way to handle it.