ItemBuildTask
-Added an ItemBuildTask to the MainGame, to demo the drone making a
battery. Currently, all it does is, when it gets to a particular power,
rather than charging, it makes a battery, which gives it a super large
amount of power. I will be changing this shortly to increasing the max
power of the drone. For now, it just supercharges it.

-I added a drone out in the wilderness, so that it would die, as
demonstration

-I changed up the movement algorithm very slightly, so that when the
pathfinding algorithm is called, it also gets passed the drones
movementAbility. I defaulted the movementAbility to a very high number,
so the drones can always move regardless, but that's easy to change, for
various movement requirements for mountain, sand, etc. That will also
allow for items that change movement ability.