Implemented a RepairTask
-As the title says, I implemented a RepairTask for the drones, much like
charging.

-As an aside, the game is running much more slowly with the comparator
in path finding; we may want to check that out.

-Additionally, it seems the MainGame (or MainGUI, whichever does it now)
is not assigning the PowerPlant builder task. This is fine in general,
since when we are actually playing, we'll have the UI, but right now it
means the drones wander off forever, power is a bit funky with no power
plant, and so on.